[src] Create sampler.py for AudioSet
